Add restrictions
Add restrictions to models and DTOs

+++ TechnicalTest.Infraestructure/Services/Candidates/Commands/CreateCandidateExperienceCommand.cs
@@ -2,6 +2,6 @@
 using TechnicalTest.Domain.DTOs;
 
 namespace TechnicalTest.Infraestructure.Services.Candidates.Commands;
-public record CreateCandidateExperienceCommand(int IdCandidate, string Company, string Job, string Description, int Salary, DateTime BeginDate, DateTime? EndDate) 
+public record CreateCandidateExperienceCommand(int IdCandidate, string Company, string Job, string Description, decimal Salary, DateTime BeginDate, DateTime? EndDate) 
     : IRequest<CandidateExperienceDto>;
